[feature] Make sure to correctly resolve binary resources from the database as well as XML resources
